move client settings db query to separate script file, remove unused function
